requirements.txt
asgiref==3.11.0
autopep8==2.3.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-silk==0.5.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.0
drf-yasg==1.21.11
drf-spectacular==0.27.2
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pycodestyle==2.14.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
pyotp==2.9.0
qrcode[pil]==8.0
cryptography==44.0.0
google-auth==2.38.0
celery==5.4.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
mediapipe==0.10.33
opencv-python-headless==4.8.0.76
numpy==1.26.4
boto3==1.40.61
openpyxl==3.1.5
xhtml2pdf==0.2.17
playwright==1.54.0
